/ 0 safety

+++ tools/restore.cpp
@@ -95,7 +95,7 @@ public:
         
         long num = 0;
         
-        int msgDelay = 1000 * ( mmf.length() / ( 1024 * 1024 * 400 ) );
+        int msgDelay = (int)(1000 * ( mmf.length() / ( 1024.0 * 1024 * 400 ) ));
         while ( read < mmf.length() ) {
             BSONObj o( data );
             
@@ -105,7 +105,7 @@ public:
             data += o.objsize();
 
             if ( ! ( ++num % msgDelay ) )
-                out() << "read " << read << "/" << mmf.length() << " bytes so far. (" << (int)(read * 100 / mmf.length()) << "%) " << num << " objects" << endl;
+                out() << "read " << read << "/" << mmf.length() << " bytes so far. (" << (int)( (read * 100) / mmf.length()) << "%) " << num << " objects" << endl;
         }
         
         out() << "\t "  << num << " objects" << endl;
